Update patch set 3
Patch Set 3:

Thanks for the input!

Recently there has been a change that was accepted (disable touchpad on soft lock) that made the first differentiation between buttons and a touchpad: the change came with a "touchpad.c".

I think I should start looking in that direction to do this thing properly.

Like you said, it would be nice if the code for gestures (including this scrollwheel) was in some non-target-specific file. Otherwise, it would have to be re-invented for every target with a touch-screen or touchpad.

On the other hand, this would have to be a pretty versatile bit of code, because I imagine the amount and type of data coming from different models of touchscreen/touchpad hardware will be quite different (different coordinates, refresh rates, maybe hardware gesture support, etc). Nevertheless, this should be doable as long as the gestures are all recognised in software. Any minor differences will just have to be normalised out by the gesture code. I can only test for real on the Fuze+ though, since that's the only one I have.

Would you agree with the following structure?

- A new target-specific driver file, via say touchpad-target.h (if there is a HAS_TOUCHPAD) is responsible for getting generic touchpad data from the touchpad hardware. (say x-y absolute and relative coordinates, number of fingers, and the touchpad resolution and refresh rate among others). This is stored in a standard struct that can be used for all targets with a touchpad.

- A non-target-specific file, say gestures.c, will get touchpad data from the aforementioned driver (or from a touchscreen driver if it's a screen instead of a pad). It will have to deal with the various types of touchscreens and touchpads available, and do correct gesture detection regardless of the resolution, refresh rate, etc of the touchscreen/touchpad. It could detect taps, swipes and a virtual scrollwheel, each with their own identifiers. This may be the biggest challenge.

- the target-specific button driver (in the case of the Fuze+ this is button-fuzeplus.c) will include gestures.h and convert the detected gestures into button presses merged with the actual hardware buttons, which are used to control Rockbox via button.h/button.c as it is now.

- plugin.h will include gestures.h and touchpad-target.h, and add functions/pointers into the plugin API so that plugins have direct access to both the raw touchpad data and the gesture detection. I would like this, because I'd like to make cool plugins that do stuff with the touchpad.

This sounds like a neat way to do it. I still need to look into the current touchscreen code, because I haven't done that at all yet. The touchpad.c would have to be a kind of "parrallel" to touchscreen.c (is there a touchscreen.c?).

The only thing I don't like about it, is that it would generalize the gesture detection to a point that it can only be software-based. Some touchpads (even the Fuze+ one) have limited support for hardware-based gesture detection, but that would be hard to implement in this structure.

Please let me know what you think - I'll wait for some more opinions because I don't want to waste time doing the wrong thing.

In any case, thanks!

Patch-set: 3
Reviewer: Gerrit User 1000285 <1000285@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0