Update patch set 1
Patch Set 1:

It does add several FRACMULs, so the features in this patch come with a definitive performance expense. There are several components of the patch which can be valued against the corresponding performance hit.  At this point in the code development, subjective listen tests are of equal importance to benchmarks, as both criteria are necessary to evaluate the value of the changes. I can submit objective (technical) data such as graphs, equations, etc., if requested.

Patch-set: 1
Reviewer: Gerrit User 1000280 <1000280@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0