ALSA driver improvements and recording support
This patch aims to improve the ALSA driver and enabling
the recording facility. This opens the possibility to
have recording on SIMULATOR and in other hosted devices
that relies on the alsa wrapper like ypr0 and ypr1.
No SDL changes for the moment.
This is still WIP, code is dirty and there are problems when
switching back to playback mode due to bad initialization sequence
I only keep here the code because I frequently change machine.

Change-Id: Ice2eaa825aef62b48b70d205bd493693f082e0e4