pcm-alsa: Add new PCM backend: async thread.
With this backend a thread is spawned that performs blocking write in a loop.
This is a different approach to the periodic writes via async notification (
signal handler) or tick task and should hopefully be the most portable backend.

Also increase the period size a bit which seems to work better on this laptop.

Change-Id: I8a570004967ca7ee5f4df782aa0f84dc77a221db