Update patch set 1
Patch Set 1:

Yea, but the threads I'm working on will have a different implementation. It's really implementation specific, the reason that thread-sdl.c looks the same is simply that it tries to emulate the original threads. I think in this case the duplicated amount of code is so tiny that it doesn't hurt.

Patch-set: 1
Reviewer: Gerrit User 1000004 <1000004@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0