Update patch set 12
Patch Set 12:

Giving the touchpad settings a submenu would build the way for future touchpad settings (if there can be any), but the gigabeatfx has one setting entry for it's touchpad presently.  This means it would look silly to have a touchpad submenu with one entry in it for the gigabeatfx.

To work around this we can define that the fuzeplus has a touchpad submenu (for it's two settings), while the gigabeatfx build does not (it has one), this only means a little more work.
OR
Let the gigabeat have a touchpad submenu with one setting in it for now, in hopes that one day a second may be added (maybe a sleep timer for powering off the pad to save power?)
OR
Leave out the submenu for now until more touchpad settings are added and then add a tuochpad section submenu when there would be more than one by default.

Thoughts?

Patch-set: 12