buflib: Add crc field protecting buflib cookie integrity
This should catch the case of buffer overflow which results
in corrupted cookie of next allocation. The check is performed
on move_block() so it may be a bit late.

There is buflib_check_valid() provided which checks the integrity
of all cookies for given context. The intention was to call it
on context switch to catch problems earlier (maby for DEBUG
build?).

Change-Id: I999d4576084592394e3dbd3bdf0f32935ff5f601