Update patch set 1
Patch Set 1:

The current code has fairly low overhead as the check is triggered only on move_block() (so rarely) and crc is calculated only over first 12bytes of every alloc in given buflib context.

The check on every context switch would be much more expensive BUT the benefit is that it should catch the thread which actually corrupted the state and maybe produce meaningful backtrace.

We could switch crc implementation to utilize full lookup table to speed up the calculation at the cost of bigger ram usage.

Patch-set: 1