Update patch set 1
Patch Set 1:

(1 comment)

Is this right that this only affects when looking in /.rockbox/cue, such that if /.rockbox/cue/track1.cue is not found that it'll look for /.rockbox/cue/track1.mp3.cue? It should not affect searchs in e.g. /Music/Albums/xx/? Your commit message is a bit unclear.

Patch-set: 1