Update patch set 2
Patch Set 2:

Just to be clear, this patch was intended as patch for people wanting to use physical keys on the fuze+ since it is a popular request. I tried to make it as clean as possible but there is some work to do to make it work on other targets (currently it breaks other targets).
You are right that we should probably have several states: not locked, locked, partial lock (whatever it is called). Feel free to pick up on this and publish an improved version.

Patch-set: 2
Reviewer: Gerrit User 1000040 <1000040@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0