Update patch set 2
Patch Set 2:

I don't see much saved from duplication in the kernel and it goes the opposite direction that I think this should go, namely the fact that it introduces more synchronization object awareness into thread.c rather than reducing or eliminating it. I have done work that removes all such awareness from the thread.c and there's no need to pass a corelock or wait queue pointer to it at all. Wait queues are handled entirely in kernel.c and thread.c knows nothing of it, simplifying thread.c in the process with a line or two more code in each kernel call. It's in my own private branch for preemptive threads on my computer but it could be ported over without much trouble. It has other other design changes as well but it still cooperative in operation unless preemption is enabled at run time.

The objective of this work is to provide a different implementation of block_thread_switch/_w_tmo() (using host-native pthreads) in the the context of a librockbox library. For this I wanted to abstract the method of blocking the current thread, such that an alternative implementation can be used.

I'm happy to hear about a different way of doing it. What can you suggest? I'd be even more happy to see your WIP code :)

Patch-set: 2