Update patch set 3
Patch Set 3:

Precisely because there is only the handle write index that has to be incremented. This change was done for preemption/parallel compatibility and I'm sharing it as an improvement but it also simplifies the code and saves about 400 bytes. In the case of tail trimming on load, it's also only one variable to adjust. It uses the same basic buffering window design (where you track the first, current and last bytes buffered of the file) that MPEGPlayer uses to be able to run on multiple processors. The only difference between this and a functionally parallel version (outside of cache coherency issues in some hardware) is the absence of multi-reader, single-writer locks. The buffering thread doesn't even need to synchonize during buffering because using only one variable for each end makes it into a non-blocking queue.

Patch-set: 3