Update patch set 3
Patch Set 3:

(1 comment)

Okay, if it makes it more preemption friendly and even reduces size too I'm all for it. It just worried me that determining the fill level of the ring buffer is not atomic (it wasn't before either), which doesn't seem preemption-safe to me.

PS: Wasn't buffering already pretty friendly for preemption due to using the separate thread and message passing for most of its operation? I used it in a pthread environment just fine.

Patch-set: 3