Update patch set 2
Patch Set 2:

That didn't answer my question.

As to your idea: I think that was the original idea of the apps/firmware split. But if you talk about real separation through a syscall interface then I think it adds unnecessary overhead for native builds. This is not related to this task though

Patch-set: 2