Update patch set 3
Patch Set 3:

The only thing that needs a proper and accurate fill level is the buffering thread, which is the only one that can actually change it anyway. For anything preemptive/parallel, there is also the matter of having to reference count direct pointers into the buffer and release them after use so they aren't moved while in use (since yielding becomes irrelevant). The scheme I had planned makes the buffering code safe but due to mrsw lock usage in my branch, any given handle is single-threaded or requires the caller to keep access serialized. This makes buffer data calls largely lock-free for readers unless it's being shrunk or the linked list is being altered. There is no multithreaded use of a given handle exept for the audio handle, which has proper cooperation. So, this patch is just the conversion with a few lines stripped away.

Patch-set: 3