Update patch set 2
Patch Set 2:

Let me paste my earlier comment (might have gone unnoticed because of my C&P failure). The question was if you have a better idea to do this.

The objective of this work is to provide a different implementation of block_thread_switch/_w_tmo() (using host-native pthreads) in the the context of a librockbox library. For this I wanted to abstract the method of blocking the current thread, such that an alternative implementation can be used.

I'm happy to hear about a different way of doing it. What can you suggest? I'd be even more happy to see your WIP code :)

Patch-set: 2
Reviewer: Gerrit User 1000004 <1000004@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0