Update patch set 2
Patch Set 2:

I have begun reviewing the keymap and I'm concerned with the fact that in many if not all screens, you put "menu" as "ok" and play/pause as either escape or got to WPS. I don't find it very logical since you naturally tend to use up/down to move and then play/pause to select. The fact that menu is not a menu further adds to the confusion. What do you think ?

I still need to go through the plugins and test on device.

Patch-set: 2