Update patch set 2
Patch Set 2:

Just to be clear, I'm fine with the overloading of functions on the buttons.
I understand your motivation to let the play/pause control playback but I think that's not the best choice. In most screens, you just don't care about being able to control it because these are very temporary screens (like pitchscreen, quickscreen, settings, ...) and in those, mapping menu to actually menu and play/pause to select is more natural. On the fuze you have a dedicated play/pause button and a select one so the situation is very different.

Patch-set: 2
Label: Code-Review=0