Sansa View: replace real bootloader with a simple test code.
After placing the firmware.mi4 file in the root dir of the player in UMS mode of
the OF, Sansa should do stupid blinking with the backlight and buttonlight
alternately. Recovering from this state is possible through the recovery mode
(see Wiki), by putting an original copy of the firmware.mi4.

Change-Id: Ia913442b97e8c405f55c4676b9a2bf0b1b1d05d6