Update patch set 1
Patch Set 1:

I have no idea how to bend the pattern, besides using a set of clones, but it slaughters my computer, which is not really that slow.
Please don't expect me to do more DAP drawing ;) Drawing the DAP in my performance is a dead slow process, it takes days and days.

Patch-set: 1
Label: Code-Review=0