Update patch set 2
Patch Set 2:

Hi,

Well #383 seems like it is trying to do a lot more. I think the build failures can be fixed with some simple ifdefs. My proposal is to fix the failing builds with something like this:

--- a/apps/plugins/lua/luaconf.h
+++ b/apps/plugins/lua/luaconf.h
@@ -371,11 +371,12 @@
-#ifndef SIMULATOR
-typedef void FILE;
+#if defined SIMULATOR || defined SDLAPP || defined ANDROID || defined SAMSUNG_YPR1
+#include <stdio.h>
+#else
+typedef int FILE;
 #endif

i.e. just include <stdio.h> where it is supported, else add a typedef for FILE. I've built the failing platforms with this and it fixes the yellows too.

If you can think of a better solution that is less ugly (I don't like hardcoding several platforms there), that doesn't require rewriting a lot of code I'm all ears. What bites here is that FILE is not actually used! we just need it to hide the types so a lot of core lua code can remain unaltered. If only there was something like an "if not typedef'd".

Cheers,

Richard

Patch-set: 2