Vibe500: improved scrollstrip driver
* a short tap in the upper/lower area of the strip generates
  single up/down-button events
* fast moves of the finger produce an additional accelleration effect
* introduces an "afterscroll" feature
* speed of strip and afterscroll is configurable:
  Settings > General settings > System > Scrollstrip speed/afterscroll
* strip can be set to "no-slide"-mode, where the upper/lower
  area of the strip is treated like ordinary buttons.
  (Sometimes very usefull, e.g. in games.)
* fix: "HAVE_LCD_FLIP" is now working

The code moves scrollstrip handling from the target specific
to a higher firmware level. This has the advantage that the
code can be reused easily for other targets.

Alas there's one shortcoming at the moment: with the highest
afterscroll setting, the player sometimes feel to "hang" at
the bottom (or top) of a list for a few seconds. This happens
when you are already near the end of the list: scrolling
stops, but as long as the player is fireing it's button-events
other buttons like "left" or "menu" are ignored.
Of course the afterscroll can be interrupted at any time by
touching the strip, but as you don't see what's going on,
the player just feels lagging.


Change-Id: Ic6b11243f354f2f973be247b578056eaf49f7f08